Add new object and user id's to batch creation and fix imports
