+++ adam/job.py
@@ -2,6 +2,7 @@
     job.py
 """
 import urllib
+import datetime
 from dateutil import parser as dateparser
 from adam import Project
